feat: add a link to homepage on the lasagna page
Added a link at the top of the page linking to the homepage.

The purpose of this link is to allow easy navigation to users back to the homepage.

This could be placed either at the top or bottom. I felt it was better to place it at the top.

It seemed logical to place this link before the main heading and not inside it. This is because any content under a heading must be logically in sync with it.