feat: list the ingredients for lasagna recipe
Ingredients list is an unordered list, meaning the order of the items do not matter.

Created an unordered list of each ingredient item.

To create fractions, instead of copying and pasting the bare text, I used `sup` `sub` HTML elements for the numerator and denominator values respectively. I used `&frasl;` to render the fraction lines.
The solution was taken from the following link on stackoverflow: https://stackoverflow.com/a/7526005/23388964